Made the project version aware.
Learned and added code to CMakeLists.txt to track and handle version numbering, and generally adjusted the file to be a bit more flexible in case of future expansion.

I then created the auto generated logger_version.h which contains preprocesser defines for the program's version number, just in case the program itself needs to know.

Lastly, cleaned up the TODO: comments and decided on a final feature set that would declare the logger to be at version 1.0.