Major overhaul and improvements to the build script.
Added a small copyright notice.

Created some general purpose helper CMake scripts that can be used by other projects.

The first helper makes a header file which allows any program using the library to be aware of the library's current version.

The second helper makes CMake configuration files for end-users to install. This allows IDEs to know the current version of the library and any special build options/variables (currently none).

Fleshed out a checklist (sort of a changelog) of all the features currently available and which features should be included next. There are 9 major features, 5 done. Thus the current
version of the project is 0.55.

Created variables to hold each part of the version number to keep things DRY.

Created variables to hold lists of source files. This keeps things things all in one place.

Added the ability to package up complete, self-contained builds of the library. The builds go into a libname-M.m.p.t-Build directory that contains the static library, header files, the test
application, and CMake configuration files. Unfortunately this step could not be automated (as far as I can tell) as part of the build process, but it's not hard to do. Simply run `cmake
--install <build-folder>` from the project root directory.

Finally, updated .gitignore to include the generated header file and the installable builds.