Added thread safety, and an appropriate stress test.
Modified the CMake build options to compile and link as a multithreaded library and test application. Then created a stress test that spawned numerous threads and forced instances of the
logger to print from those threads to shared resources as much as possible in an attempt to break the output streams. I then adjusted the logger itself to handle that properly.