Added tests that make logger fail.
I began working on making the logger tee stream output, but while I was thinking about how to store the extra output stream references I realized that something could happen to one or more
of the streams -- outside -- of the logger since it doesn't take ownership of them and only holds a reference. Curious as to what would happen I side-tracked and created the badStreamTest
and the deadStreamTest functions to try and make the tester app crash. The deadStreamTest succeeded at creating a seg-fault. Hooray! I still don't know what (if anything) I am/should do
about this design.