Added feature to tee logger streams.
After creating a logger with an initial stream you can now add more stream references that it'll store within an internal vector. It will then output each log message to each of the streams
it is aware of (assuming those streams don't have problems or get deleted).

At this point the logger is plenty usable in other projects! It could definitely use some enhancements (like the much desired log appending and rotation), but it's safe enough, and feature
full enough that it can be used as is.

NOTE: I also briefly explored the idea of storing stream references as some sort of smart pointer type (though this isn't reflected in the code). Weak pointer may work well to ensure that
Logger is using a stream that actually exists (thus preventing seg-faults), but still wouldn't alleviate problems with the streams themselves (such as a closed file stream). Using a weak
pointer would also create difficulties in using std::cout as that can't be dynamically created. The end-user would have to dynamically create an ostream object stored in a shared pointer,
they would then have to steal the address of std::cout's read buffer and assign it to the ostream they created, then they'd have to give that ostream to Logger. It's enough work that I'd
want to provide a convenience function for users so they could simply "get a terminal logger".