Improved Documentation.
Studied more about how to put Doxygen to good use while adding even more commenting to the code. I've grouped some things and mildly tweaked doxygen.conf. I also realized that I had a
function documented as returning true on success, but the function can't fail and I didn't know that when I first drafted the interface. That fuction was just always returning a set boolean
value, so I changed it to void and updated the docs to reflext that.