Found and fixed a compilation flaw.
My main motivation and intent for making this library was to embed it into a different library. Turns out that I need to set a compilation flag called -fPIC to generate "Position
Independent Code" or it won't work. It'll still be compiled into executable applications just fine, but not other libraries.

The problem is that the default setup for compiling a static library is simply to make an archive file of one or more raw object files. This is too rigid for shared libraries which need to
be able to tear apart a symbol table and integrate its pieces into its own symbol table. The PIC flag makes this possible, and thus the library more versatile in its uses.